+++ packages/pdfx/example/lib/pinch.dart
@@ -20,7 +20,7 @@ class _PinchPageState extends State<PinchPage> {
       // document: PdfDocument.openAsset('assets/hello.pdf'),
       document: PdfDocument.openData(
         InternetFile.get(
-          'http://www.africau.edu/images/default/sample.pdf',
+          'https://api.codetabs.com/v1/proxy/?quest=http://www.africau.edu/images/default/sample.pdf',
         ),
       ),
       initialPage: _initialPage,